[#84] feat : ItemController와 PlaceRankController 분리
- GET /app/item/place/top API를 PlaceRankCcontroller로 이동
- POST /app/item/place API 추가.
